pyproject.toml
[project]
name = "langchain-demo"
version = "0.1.0"
description = ""
authors = [{ name = "Jonathan Davies" }]
readme = "README.md"
requires-python = ">=3.13,<3.14"
dependencies = [
    "langchain[google-genai] (>=1.1.0,<2.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "flask (>=3.1.2,<4.0.0)",
    "ruff (>=0.14.6,<0.15.0)",
    "mypy (>=1.18.2,<2.0.0)",
    "langchain-text-splitters (>=1.0.0,<2.0.0)",
    "langchain-community (>=0.4.1,<0.5.0)",
    "bs4 (>=0.0.2,<0.0.3)",
    "langchain-core (>=1.1.0,<2.0.0)",
]
scripts = {print_hello = "langchain_demo.hello:print_hello"}

[tool.ruff]
line-length = 88
indent-width = 4

[tool.ruff.lint]
# Enable Pyflakes (`F`) and a subset of the pycodestyle (`E`)  codes by default.
# Unlike Flake8, Ruff doesn't enable pycodestyle warnings (`W`) or
# McCabe complexity (`C901`) by default.
select = ["E4", "E7", "E9", "F"]
ignore = []

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

# Enable auto-formatting of code examples in docstrings. Markdown,
# reStructuredText code/literal blocks and doctests are all supported.
#
# This is currently disabled by default, but it is planned for this
# to be opt-out in the future.
docstring-code-format = false

# Set the line length limit used when formatting code snippets in
# docstrings.
#
# This only has an effect when the `docstring-code-format` setting is
# enabled.
docstring-code-line-length = "dynamic"

[tool.ruff.lint.pydocstyle]
convention = "google" # Accepts: "google", "numpy", or "pep257".

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
